Enhance file-loader configuration to be able to handle nested paths and favicon output
Previously, the file-loader's `name` field outputted every asset to `dist/static/assets/`, regardless of the actual directory the file was in. This means that e.g. `assets/images/my-image.png` and `assets/fonts/my-font.woff` would both be outputted to `dist/static/assets/`. While this doesn't pose an issue for dist functionality, it does result in a less organized build output and breaks the IDE's ability to follow paths between src and dist because the folder structures are different.

This improvement also allows the file-loader to handle moving favicon files to `dist/` so that CopyWebpackPlugin doesn't have to do that itself. This prevents duplicating favicon files in both `dist/favicon.png` and `static/assets/favicon-hash.png`.