Add support for circular references
To handle circular references, a WeakMap was used, which uses objects as keys and anything as the value.

Specifically, by setting the original `obj` as the key and the `copy` as the value, the copy would be returned whenever a circular reference is made to said `obj`; this frees any pointers within `copy` from pointing to anything in `obj`.

Also, add tests specifically for checking altered fields in the circular reference within the `copy` such that any changes in either root-level or nested circular fields would not carry between the original `obj` and the new `copy`.

Finally, use the `arguments` keyword instead of a default parameter value in order to keep the function signature clean of any outside intervention by the user.