Add sample component to show how to handle SVG tween/interpolation animations
In order for an SVG to have a tween/interpolation animation (where it dynamically generates the intermediary frames to get from one child's display to the next), the `<animate>` tag requires both a `from` and a `to` attribute. In order to create the `from` value, the components previous value for e.g. `<path.d>` needs to be remembered, so a ref is used which tracks the previous value without causing double renders (which would happen if it were in state).

Also, since React reuses HTML elements rather than unmounting/remounting them, changing props/re-rendering the component doesn't trigger the animation from `<animate>`. This can be fixed by getting the HTML element from the `<animate.ref>` field and using it to call the native DOM function `beginElement()`, which resets the animation start time to now, resulting in a fresh animation.