Make the initial "grow" animation a bit slower than the tween animations
Since growing the line from (0,0) to the final selected line is a much bigger change than from a pre-existing line to a new line, animating it in 0.15 seconds is significantly jarring. Increasing the animation time results in a smoother experience for an animation that has such bigger differences from start to end.