Add custom loader for SVGs
To allow SVGs to be importable as React components (e.g. to allow forwarding props down to the SVG), '@svgr/webpack' was added. But, to also allow them to be importable as URLs (e.g. CSS 'background: url()' or <img src='svg-url' />), 'file-loader' was also necessary, mainly because 'type: asset/resource' doesn't seem to play nicely with other loaders. The final result is that the URL is the default export and the React component can be imported via named ReactComponent field. Finally, allow refs to be forwarded to the root <svg> via @svgr/webpack config option