Add source maps for huge usability improvement during development
[Source maps](https://developer.mozilla.org/en-US/docs/Tools/Debugger/How_to/Use_a_source_map) are used to show which source file is associated with which build output file. These are super useful during development because it makes the links to files displayed in the browser's dev tools (e.g. in Network pane showing what file made the request, in CSS style pane showing where the style is defined, etc.) point to the original source files rather than the bundled/transpiled/minified output files. Those source files are easier to read/interpret than the bundled output (obviously) AND they be used for debugging (so you're not forced to use bundled output for debugging).

To generate source maps automatically, we configure it in the [webpackConfig.devtools](https://webpack.js.org/configuration/devtool/) field. There are lots of different types of source maps, but this particular choice is the most helpful during development as it shows the absolute original source code rather than transpiled, minified, bundled, or polyfilled code. This option comes at the cost of a slightly longer build/rebuild time, but that's worth it for the benefits.

Finally, we don't want source maps in our production output code, so disable it during build. Note that since it's disabled during build, that means we don't have to pass it down to the `webpackConfig.optimization.minimizer[]` entries (granted, we likely wouldn't have to anyway since their default config objects are generated from the webpack config object, but that may depend on the plugin and/or its version).