Replace hard-coded paths with 'Paths' usage
Not only does this replace the magic strings for paths all across the config files, but it also results in a single source of truth to avoid confusion later for which path means what. Consequentially, this also makes the app more flexible for users because changing the names of these directories later on is substantially easier if they wish.

Note: This doesn't work for tsconfig.json because [the TypeScript devs are stubborn](https://github.com/microsoft/TypeScript/issues/25271), so those paths would have to be updated manually if they're changed in `Paths`.