Return 'null' for null SCSS values
In JS, the imported value from SCSS is always an object as a default export (it isn't spreadable in the import statement). As such, a non-existent key will always be `undefined`, i.e.

```javascript
import MyStyles from './styles.scss';
console.log(MyStyles.nonExistentKey); // === undefined
```

This means the previous logic of keeping SCSS null values as null worked fine for root-level exports due to the fact that SCSS doesn't even export null values at all. However, for nested null values (i.e. in a list or map), SCSS' complete erasure of the value actually breaks the list/map parsing (e.g. the generated string is `{"nullKey": , "otherKey": "something"}` instead of `{"nullKey": null, "otherKey": "something"}`). Thus, add in `'null'` as the value so that `JSON.parse(MyStyles.myMapJsonString)` doesn't fail.

Note: `'null'` has to be put in quotations to make it a string in SCSS so it isn't erased by the SCSS export logic, but we don't use `stringify()` because that would result in `'"null"'` which would be parsed by JavaScript as `'null'` instead of `null`.

Also, do some minor comment updates to reflect that only strings/booleans are fine as-is and to keep the order of variable type processing consistent.