Remove unnecessary CLI arg processing
Turns out that manual CLI arg processing isn't needed if the user uses the correct CLI args for the command that's being run under the hood, even if that command is wrapped by our own configuration. Thus, remove the manual configuration for `--coverage`, but leave it commented so the format can be seen/understood for future custom CLI arg parsing.