Update babel.config to inject core-js by usages
This helps simplify our config by not having to import (or in the previous case, subtle-import by using it as an entry file) core-js in each of the app's entry files in order for the polyfills to be applied. In effect, this means that each separate output file will be able to function as a standalone module without reliance on its parent having imported core-js already.

In particular, this will be useful for:

1. The case where multiple entry points will be used in a multi-page app (instead of a SPA) because it allows each page to have their own customized configurations that aren't restricted by a dependency on importing core-js or adding it as entry file.
2. Tree-shaking will reliably output standalone polyfilled code rather than code that needs to be polyfilled by the parent.

See: https://babeljs.io/docs/en/babel-preset-env#usebuiltins