Use standard URL-path routing instead of hash routing
Since both `HashRouter` and `BrowserRouter` use the `history` API, they both use client-side routing, not server-side routing. As such, they are equivalent in functionality other than the route having a preceding `#` before it or not in the URL.

Since `#` is also used to auto-navigate to the element that has a matching `id` on it, using `HashRouter` conflicts with this functionality. Thus, change the router to use basic/standard `/` routing.