Minor: Replace all `window` references with `self`
`self` has been defined since (almost) the beginning of time in the browser as a proxy/pointer to `window`. However, `window` is not defined in the `(Service|Web)Worker` environment. While none of these files are used in the Worker environment currently, change them to use `self` anyway in case they will be used by a Worker in the future.