Rename tests to use `.test.ext` format instead of `.spec.ext`
'spec' stands for 'specification' which better represents what goes in the `it()` description, in reality, the test files will more often than not have a mixture of these plus non-spec tests in them. For consistency in file names and because of the mixture of tests, just name all files `.test.ext`. This also has a slight additional benefit in that the file's purpose is immediately apparent (vs if some sort of test util file didn't have the extension prefix) and prevents implying that some tests should be moved from one extension prefix file to another.