Move Jest `setupFiles` to `setupFilesAfterEnv` so that `jest` namespace is defined
`setupFiles` are run before the jest environment is setup (i.e. before `jest`, polyfills, etc. are added in), whereas `setupFilesAfterEnv` is run after all that setup is done. As such, we can't call functions like `jest.beforeEach()` if a file is in `setupFiles` but we can if it's in `setupFilesAfterEnv`.

Relatedly, reset `(local|session)Storage` before each test to ensure their values don't bleed through from test to test.