Dynamically parse tsconfig import aliases for all other configs to use
This results in a single source of truth for where import aliases are defined, so the code isn't repeated over and over. As such, it means that each type of config file (Webpack, Jest, ESLint, etc.) can format the aliases as they wish without having to hard-code the alias-pathMatch mappings themselves. It also means we don't need to worry about forgetting to update all the config files if any alias changes are made or additional aliases are added.