Ensure IDEs always autocomplete imports in tests using aliases
At some point during the many improvements made along the way, import autocompletion in test files stopped using import aliases. By adding a new tsconfig.json file in that directory, the IDE correctly reads the import aliases from the root-level tsconfig.json as it should.

Theoretically, this is the same issue as what happened when the root-level tsconfig.json was in config/ instead of root; as such, it can be fixed in the same way - by putting a tsconfig.json file in the respective directory where you want to ensure certain behaviors occur and to enforce those behaviors in IDEs.