Bug fix: Don't default functional components `props` to empty object
Unlike normal functions, React defaults `props` to an empty object instead of leaving it `undefined`. As such, we don't need to default `props` to an empty object like we have to do for normal functions.

This is especially important for TypeScript functional components. If you want to spread the `props` object fields in the function signature and add default values for them, then you would have to *duplicate* all the non-optional values in the object after the `=`. And, even if you add a default value for them in the object after `=`, they *still won't default the value* because components without any props passed in are still given `props = {}`.
i.e.

```tsx
interface MyCompProps {
    a: string;
    b?: number;
    c?: string;
}

/* Correct */
function MyComp({ a = 'A', b = 3, c = 'C' }: MyCompProps) {
    return (
        <div>a: {a}, b: {b}, c: {c}</div>
    );
}

/* Incorrect */
// Have to duplicate the values in TWO places:
// The ones on the left actually fill in the `props` values.
// The ones on the right are there just to make it match the correct interface but don't actually do anything!
function MyComp({ a = 'A', b = 3, c = 'C' }: MyCompProps = { a: 'A' }) {
    return (
        <div>a: {a}, b: {b}, c: {c}</div>
    );
}

/* Both produce the following */

// No props
<MyComp /> // ERROR: `a` must be passed in
// Some props
<MyComp a={'hello'} /> // <div>a: hello, b: 3, c: C
```