Minor: Spread props for functional components in their signature
After the recent commit removing the default value of `props = {}`, this helps avoid the problem occurring in the future by spreading the props in the function's signature instead of using the `props` object.

Also, this allows you to add default values for each prop as you wish without having to rewrite the template code.