Move JSDoc for Jest util function to overload signature
TypeScript doesn't expose the function signature/JSDoc for the implementation of overloaded functions, it only exposes the signature/JSDoc of the function overload signatures themselves. Thus, move the actual, helpful JSDoc to the overloaded function signature so it's what appears in IDEs instead of the dev notes about how/why the overload signature was added.