Make ESLint understand JSON file imports
Now that we added Node support for importing `.json` files directly, we should add it as a valid import extension; otherwise, ESLint complains that 'file.json could not be resolved'. However, don't add it to the top-level `extensions` array because we don't want ESLint to lint those files, we only want it to recognize that they're valid.