Add enhanced decorator syntax for TypeScript and new `types/` dir for global types
The `types/` dir is useful for global types/type modifications that don't have a `@types/` entry (as described). Likewise, with the uncertainty of decorator syntax from the previous commit, this at least provides more information within decorators to use while we wait for the better/enhanced TC39 decorator syntax to be finalized.

Respectively, add the new `Reflect` type-modifying library as a global type.