Minor: Remove extra spaces/newlines introduced by removing comments; Keep one newline for readability
Previously, the comment deletion logic deleted the newline separating the line the comment was on and the next line due to the `//` comment deletion regex including the final `\n`. This fixes that by keeping that newline when deleting the comment while also deleting any additional whitespace/newline characters from the ends of lines.