Speed up ESLint by 10-20% by using only TypeScript's ESLint parser
Since TypeScript files still have to be processed and since both Babel/TypeScript ESLint parsers are the same other than added support for typedefs, replace Babel with TypeScript so we're not spawning multiple parsers to handle files that co-op with each other. Speed appears to improve by an average of 15% by doing so, but the max/min ranges a bit wider; regardless, this is clearly the way to go.

Likewise, uninstall the no-longer-used `@babel/eslint-parser` dependency.