Bug fix: Fix naming conflict between '<project>/' and '<root>/'
While this is ultimately a bug with `tsconfig-paths` since it worked fine in source code, in order to use imports relative to the project-root within `ts-node` scripts, we must avoid import-name conflicts with '/path/to/file' since it attempts to read those as imports from the (Unix-based) file system itself. Thus, replace the '/' `paths` alias with '~/' to allow clear distinction between the two.

Also, remove completed TODO from ReadMe.