Upgrade `@svgr/webpack` and configs to maintain named/default exports
After `(file|url|raw)-loader` was deprecated by Webpack in favor for asset modules, `@svgr/webpack` failed to maintain the nice behavior of exporting both the asset file's URL and React component. In order to maintain this, we must specify a ton of configs to pick up the slack that SVGR left behind. Thus, upgrade the dependency, add the ridiculous amount of configs necessary to patch their inequities, and update Jest mocks to match these changes.