Minor: Simplify test's bug fix by using correct vars/order of ops
It's likely one of the reasons the tests failed previously had to do with `JSON.stringify()`, using `importedSvg` (the test-defined var, not the tested var) in `expect()` instead of `.toEqual()`, or something else. It's also clear this is only an issue in the CI so the real issue is unknown, but hopefully this gives some extra context for why it failed previously.