Bug fix: Better support multiple import alias path-matches
It was possible that, before this change, an import alias that had multiple path-matches would cause Jest, ESLint, Babel, etc. to fail. This change allows them to call the `pathMatchModifier()` on each individual path match for an alias if they want to deal with each string separately while still maintaining the original functionality if they want to deal with all import paths at once via a Regex string.